chore: start refactoring pages
