+++ src/components/footerLogo/index.module.scss
