+++ src/components/homeBanner/index.module.scss
